requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.19.5
pandas==1.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.1
pytz==2020.5
requests==2.25.1
sendgrid==6.5.0
six==1.15.0
soupsieve==2.1
starkbank-ecdsa==1.1.0
urllib3==1.26.2
